% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/british_to_american_spellings.R
\name{british_to_american_spellings}
\alias{british_to_american_spellings}
\title{Return character "dictionary" for converting British to American spellings}
\usage{
british_to_american_spellings()
}
\description{
Return character "dictionary" for converting British to American spellings.
}
