% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_stm_input.R
\name{create_stm_input}
\alias{create_stm_input}
\title{Create objects for input to STM functions}
\usage{
create_stm_input(input_dir, output_file, bounds = c(2, Inf),
  word_lengths = c(3, 24), split_docs = FALSE)
}
\arguments{
\item{input_dir}{Directory containing text files from which to create STM input.}

\item{output_file}{Full filepath, including filename, at which to save .Rdata file containing STM input objects.}

\item{bounds}{An integer vector of length 2 indicating the lower and upper bounds (inclusive) for the number of documents in which a token must appear to be included in the analysis.}

\item{word_lengths}{An integer vector of length 2 indicating lower and upper bounds (inclusive) for the number of characters a token must have to be included in the analysis.}

\item{split_docs}{Logical indicating whether documents should be split into 3,000 word sections prior to STM analysis.}
}
\description{
Create and save objects to be used as input to STM functions.
}
