% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_text_from_pdfs.R
\name{extract_text_from_pdfs}
\alias{extract_text_from_pdfs}
\title{Extract text from PDF files}
\usage{
extract_text_from_pdfs(pdf_dir, output_dir, save_Rdata = TRUE,
  save_txt_files = FALSE, language = "en")
}
\arguments{
\item{pdf_dir}{Directory containing PDF files to extract text from.}

\item{output_dir}{Directory to save output to.}

\item{save_Rdata}{Logical indicating whether to save VCorpus object named "all_reports" as an .Rdata file.}

\item{save_txt_files}{Logical indicating whether to save documents as individual text files.}

\item{language}{Language in which documents are written.}
}
\description{
Extract text from PDF files and save as VCorpus and/or in individual text files.
}
