% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ngrams.R
\name{get_ngrams}
\alias{get_ngrams}
\title{Extract and save ngrams from text}
\usage{
get_ngrams(input_dir, ngram_dir, language = "en")
}
\arguments{
\item{input_dir}{Directory containing text files to extract ngrams from.}

\item{ngram_dir}{Directory to save ngrams to.}

\item{language}{Language in which documents are written.}
}
\description{
Extract and save ngrams from text to aid in the removal of boilerplate text from documents.
}
