% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_sure_dir.R
\name{make_sure_dir}
\alias{make_sure_dir}
\title{Create directory if it doesn't already exist}
\usage{
make_sure_dir(dir)
}
\arguments{
\item{dir}{Directory to create if it doesn't exist.}
}
\description{
Remove repetitive "boilerplate" text from documents to minimize noise in the STM analysis.
}
