% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_boilerplate.R
\name{remove_boilerplate}
\alias{remove_boilerplate}
\title{Remove repetitive "boilerplate" text from documents}
\usage{
remove_boilerplate(input_dir, ngram_dir, output_dir, rep_text_dir,
  header_footer_dir, language = "en")
}
\arguments{
\item{input_dir}{Directory containing text files to extract ngrams from.}

\item{ngram_dir}{Directory in which to find ngrams.}

\item{output_dir}{Directory in which to save texts with boilerplate removed.}

\item{rep_text_dir}{Directory in which to save repetitive text for review.}

\item{header_footer_dir}{Directory in which to save header and footer text for review.}

\item{language}{Language in which documents are written.}
}
\description{
Remove repetitive "boilerplate" text from documents to minimize noise in the STM analysis.
}
