% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_named_entities.R
\name{remove_named_entities}
\alias{remove_named_entities}
\title{Remove named entities using Python module ntlk}
\usage{
remove_named_entities(input_dir, output_dir, nes_dir)
}
\arguments{
\item{input_dir}{Directory containing text files to remove named entities from.}

\item{output_dir}{Directory in which to save texts with named entities removed.}

\item{nes_dir}{Directory in which to save named entities for review.}
}
\description{
Remove named entities using Python module ntlk.
}
