% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_state_and_country_names.R
\name{remove_state_and_country_names}
\alias{remove_state_and_country_names}
\title{Remove state and country names}
\usage{
remove_state_and_country_names(input_dir, output_dir)
}
\arguments{
\item{input_dir}{Directory containing text files from which to remove state and country names.}

\item{output_dir}{Directory to which to save documents with state and country names removed.}
}
\description{
Remove a set list of countries and US states from documents.
}
