% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/substitute_documents.R
\name{substitute_documents}
\alias{substitute_documents}
\title{Extract text from PDF files}
\usage{
substitute_documents(corpus_dir, replacements_dir, replacement_scheme)
}
\arguments{
\item{corpus_dir}{Directory containing documents in individual text files.}

\item{replacements_dir}{Directory containing replacement documents.}

\item{replacement_scheme}{A named character vector in which the names are the filenames (without extensions) of documents to be replaced, and the values are the filenames (without extensions) of documents to serve as replacements.}
}
\description{
Extract text from PDF files and save as VCorpus and/or in individual text files.
}
